\name{recalls-package}
\alias{recalls-package}
\docType{package}
\title{Access U.S. Federal Government Recall Data}
\description{recalls provides access to U.S. Federal Government recall data on drugs and food from the Food and Drug Administration, car safety from the National Highway Traffic Safety Administration, and product safety recalls from the Consumer Product Safety Commission.}
%\details{}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://search.digitalgov.gov/developer/recalls.html}{API Documentation}
}
\keyword{ package }
%\seealso{}
%\examples{}

